Move metric to server-core
